third_party/blink/web_tests/fast/dom/gc-dom-tree-lifetime-expected.txt

=== Initial state ===
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
=== After clearing innerHTML ===
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
=== After clearing innerHTML and divX ===
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
=== After clearing innerHTML, divX and divY ===
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
=== After clearing innerHTML, divX, divY and divZ ===
=== Initial state ===
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
=== After clearing innerHTML ===
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
=== After clearing innerHTML and divX ===
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
=== After clearing innerHTML, divX and divY ===
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
=== After clearing innerHTML, divX, divY and divZ ===
=== Initial state ===
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
=== After clearing innerHTML ===
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
=== After clearing innerHTML and divX ===
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
=== After clearing innerHTML, divX and divY ===
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
=== After clearing innerHTML, divX, divY and divZ ===
=== Initial state ===
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
=== After clearing innerHTML ===
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
=== After clearing innerHTML and divX ===
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
=== After clearing innerHTML, divX and divY ===
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
=== After clearing innerHTML, divX, divY and divZ ===
=== Initial state ===
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
=== After clearing innerHTML ===
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
=== After clearing innerHTML and divX ===
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
=== After clearing innerHTML, divX and divY ===
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
=== After clearing innerHTML, divX, divY and divZ ===
=== Initial state ===
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
=== After clearing innerHTML ===
PASS globalDiv.id is "div2"
PASS globalDiv.parentNode.id is "div2-parent"
PASS globalDiv.firstChild.id is "div2-child"
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
=== After clearing innerHTML and divX ===
PASS globalDiv.id is "div1"
PASS globalDiv.parentNode.id is "div1-parent"
PASS globalDiv.firstChild.id is "div1-child"
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
=== After clearing innerHTML, divX and divY ===
PASS globalDiv.id is "div0"
PASS globalDiv.parentNode.id is "div0-parent"
PASS globalDiv.firstChild.id is "div0-child"
=== After clearing innerHTML, divX, divY and divZ ===
PASS successfullyParsed is true

TEST COMPLETE
